requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
discord-protos==0.0.2
discord.py-self @ git+https://github.com/dolfies/discord.py-self
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
protobuf==5.29.3
python-dotenv==1.0.1
tzdata==2025.1
tzlocal==5.2
yarl==1.18.3